+++ src/main/java/model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import java.util.List;
 import java.util.stream.Collectors;
