+++ src/test/java/model/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
